Rearrange parser function for complying with norm.
I changed the structure of the function to make them complient with the norm.
Also added the expander to the filename if it contains a $ in its name to check if it is in the env variables.
